Move TestHttpRequestData to own file
